Added missing conf-publisher binary and minor improvements
