+++ config/src/main/assembly/assembly.xml
@@ -33,8 +33,8 @@
             </includes>
         </fileSet>
         <fileSet>
-            <directory>${basedir}/src/main/bin</directory>
-            <outputDirectory>bin</outputDirectory>
+            <directory>${basedir}/src/main/scripts</directory>
+            <outputDirectory>.</outputDirectory>
             <includes/>
             <fileMode>755</fileMode>
         </fileSet>
